🧐 add intermediate analysis after cycle 82
Raw data files have been removed from the repository, because the size
has gotten too large for Git LFS. Some lower-resolution data may be
added back to the repo, or data storage may be moved to DVC in the
future.

The download.sh script references a local parflux installation on a path
outside this repository. This is not portable and needs to be changed
before merging this branch to main.